refactor(trace): share trace result types across modules
